Ahora el mainViewModel se separa de cuadradoViewModel, por lo que se crea la clase CuadradoViewModel.
